refactor: renamed chat collection to messages
